+++ domain/src/main/kotlin/team/duckie/app/android/domain/recommendation/usecase/FetchFollowingTestUseCase.kt
@@ -16,7 +16,7 @@ class FetchFollowingTestUseCase @Inject constructor(
     private val repository: RecommendationRepository,
 ) {
 
-    suspend operator fun invoke() = kotlin.runCatching {
+    suspend operator fun invoke() = runCatching {
         repository.fetchFollowingTest()
     }
 }
